+++ src/config/deserialized_account_config.rs
@@ -1,5 +1,5 @@
 use serde::Deserialize;
-use std::path::PathBuf;
+use std::{collections::HashMap, path::PathBuf};
 
 pub trait ToDeserializedBaseAccountConfig {
     fn to_base(&self) -> DeserializedBaseAccountConfig;
@@ -39,12 +39,6 @@ macro_rules! make_account_config {
             pub signature_delimiter: Option<String>,
             /// Overrides the default page size for this account.
             pub default_page_size: Option<usize>,
-            /// Overrides the inbox folder name for this account.
-            pub inbox_folder: Option<String>,
-            /// Overrides the sent folder name for this account.
-            pub sent_folder: Option<String>,
-            /// Overrides the draft folder name for this account.
-            pub draft_folder: Option<String>,
             /// Overrides the notify command for this account.
             pub notify_cmd: Option<String>,
             /// Overrides the IMAP query used to fetch new messages for this account.
@@ -75,6 +69,10 @@ macro_rules! make_account_config {
             /// Represents the command used to decrypt a message.
             pub pgp_decrypt_cmd: Option<String>,
 
+    	    /// Represents mailbox aliases.
+    	    #[serde(default)]
+    	    pub mailboxes: HashMap<String, String>,
+
 	    $(pub $element: $ty),*
 	}
 
@@ -86,9 +84,6 @@ macro_rules! make_account_config {
             	    signature: self.signature.clone(),
             	    signature_delimiter: self.signature_delimiter.clone(),
             	    default_page_size: self.default_page_size.clone(),
-            	    inbox_folder: self.inbox_folder.clone(),
-            	    sent_folder: self.sent_folder.clone(),
-            	    draft_folder: self.draft_folder.clone(),
             	    notify_cmd: self.notify_cmd.clone(),
             	    notify_query: self.notify_query.clone(),
             	    watch_cmds: self.watch_cmds.clone(),
@@ -105,6 +100,8 @@ macro_rules! make_account_config {
 
             	    pgp_encrypt_cmd: self.pgp_encrypt_cmd.clone(),
             	    pgp_decrypt_cmd: self.pgp_decrypt_cmd.clone(),
+
+		    mailboxes: self.mailboxes.clone(),
 		}
 	    }
 	}